game.py -> loop.py, game.py and commenting
- game.py was renamed loop.py

- game.py was created to house the main game logic. currently it has 2
functions __init__ (used for initialisation) and update which is called
every loop cycle to update the game

- a small amount of commenting was also added

- the modules src.map and src.board were removed
